Uploaded new version of interface file
Change: stopped display of header above graphs reading "Results for California and Quebec"; seemed redundant. (Change involved removing 3 lines of code to define the header text and display it.)

Note that this header text was not being displayed after running the model again; it would be cleared on clear_output.